You can get coordinates

+++ app/src/main/java/cheysoff/weather/City.kt
@@ -1,13 +1,13 @@
 package cheysoff.weather
 
-class City(private val name : String, private val latitude: Int, private val longitude: Int) {
+class City(private val name : String, private val latitude: Double, private val longitude: Double) {
     fun GetName() : String {
         return name
     }
-    fun GetLatitude() : Int {
+    fun GetLatitude() : Double {
         return latitude
     }
-    fun GetLongitude() : Int {
+    fun GetLongitude() : Double {
         return longitude
     }
 }
